begin code re-org and cleanup
